Revert "Merge branch 'NoAbstraction' of https://github.com/Brenstem/AsteroidsProj into NoAbstraction"
This reverts commit 2c2f1fd502723a621243b9458d31164b29b08004, reversing
changes made to 3ea98b1979cab1de2f26668dd6fc70f218d74adb.

+++ SFMLConsoleApplication/SFMLConsoleApplication/Game.h
@@ -65,7 +65,7 @@ private:
 	void handleLostCoin();
 	void handleAsteroidLoss();
 
-	int getRandomNumber(int min, int max);
+	int getRandomNumber(float min, float max);
 
 };
